move verilator specific files to their own directory
